Rename dashboard-splash-app.js to DashboardSplashApp.js.
