+++ assets/js/components/dashboard-splash/DashboardSplashApp.js
